removed pickup code. Calls to StartsWith replaced with LineBuffer
Former-commit-id: cebff44bb777b9f1ecf1285eea4818b74d24e5c0

+++ src/gmap.cpp
@@ -19,7 +19,7 @@ u32 HashString(u32 Size, const void* Key)
 u32 HashStringJenkins(u32 Size, const void* Key) 
 {
     char* K = (char*)Key;
-    u32 Length = strlen(K);
+    u32 Length = (u32)strlen(K);
     u32 I = 0;
     u32 Hash = 0;
     while (I != Length) {
@@ -33,7 +33,6 @@ u32 HashStringJenkins(u32 Size, const void* Key)
     return Hash & (Size - 1);
 }
 
-
 u32 HashInt(u32 Size, const void* Key) {
     u32 K = *(u32*)(Key);
     K = ((K >> 16) ^ K) * 0x45d9f3b;
@@ -41,6 +40,7 @@ u32 HashInt(u32 Size, const void* Key) {
     K = (K >> 16) ^ K;
     return K % (Size - 1);
 }
+
 /*
 u32 HashInt(u32 Size, const void* A)
 {
